+++ NEWS.md
@@ -1,5 +1,4 @@
-# sjstats 0.12.0.9000
-
+# sjstats 0.13.0
 ## General
 
 * Remove deprecated `get_model_pval()`.
@@ -12,6 +11,7 @@
 
 ## Changes to functions
 
+* `link_inverse()` now supports `brmsfit` and `clm`-models.
 * `var_names()` now also cleans variable names from variables modelled with the `offset()` function.
 * `icc()`, `re_var()` and `get_re_var()` now support `brmsfit`-objects (models fitted with the *brms*-package).
 * For `fun = "weighted.mean"`, `typical_value()` now checks if vector of weights is of same length as `x`.